+++ src/index.ts
@@ -14,7 +14,7 @@ export {
 export { useMotionValue } from "./value/use-motion-value"
 export { MotionValue, motionValue } from "./value"
 export { unwrapMotionValue } from "./value/utils/unwrap-value"
-export { useTransform } from "./value/use-transformed-value"
+export { useTransform } from "./value/use-transform"
 export { useViewportScroll } from "./value/use-viewport-scroll"
 export {
     AnimationControls,
